refactor: move components into ManagerComponents folder for easier manageability.
